requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic-core>=2.41.5
python-multipart==0.0.6
openai==1.3.0